// src/App/Resources/Enums/Routes.ts
export enum Routes {
    CAREER = "/career",
    CONFS = "/confs",
    CONTACT = "/contact",
    GITHUB = "https://github.com/HelviraG",
    HOME = "/",
    LINKEDIN = "https://www.linkedin.com/in/helvira-dev",
    LIVE = "/lives",
    PRESS = "/press",
    TWITTER = "https://twitter.com/helvira_g"
}